pyproject.toml
[tool.poetry]
name = "lolify-instagram"
version = "0.1.0"
description = ""
authors = ["Shub Raj Lama <hello@shubraj.com>"]
readme = "README.md"
packages = [{include = "lolify_instagram"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
